libmount: add function to reliable read mountinfo
The standard way how we read mount table is not reliable because
during the read() syscalls the table may be modified by some another
process. The changes in the table is possible to detect by poll()
event, in this case it seems better to lseek to the begin of the file
and read all again. It's expensive, but better than races...

Note that this patch only add a new function, but the function is not
integrated to libmount mountinfo parser yet. For now it's for testing
purpose only.

Reported-by: Zbigniew Jędrzejewski-Szmek <zbyszek@in.waw.pl>
Signed-off-by: Karel Zak <kzak@redhat.com>